feat(web): enhance daily plan components with new skeletons and refactor imports
- Introduced `AllPlans`, `EmptyPlans`, and `PlanHeader` components to improve the daily plan feature, providing better structure and functionality.
- Added `ActivityCalendarSkeleton` for improved loading states in the activity calendar.
- Refactored existing components to utilize the new skeletons, ensuring a consistent loading experience across the daily plan section.
- Updated imports to streamline component usage and enhance code organization.
- Removed the obsolete `ScreenCalendar` component to simplify the codebase.
