--- apps/web/core/components/activities/screen-calendar.tsx
@@ -1,7 +1,0 @@
-'use client';
-import React from 'react';
-import { ActivityCalendar } from './activity-calendar';
-
-export function ScreenCalendar() {
-	return <ActivityCalendar />;
-}
