+++ apps/web/core/components/daily-plan/index.ts
@@ -4,3 +4,7 @@ export * from '../features/daily-plan/enforce-planed-task-modal';
 export * from '../features/daily-plan/add-daily-plan-work-hours-modal';
 export * from '../features/daily-plan/suggest-daily-plan-modal';
 export * from '../features/daily-plan/add-task-estimation-hours-modal';
+export * from './all-plans';
+export * from './all-plans-modal';
+export * from './empty-plans';
+export * from './plan-header';
