+++ apps/web/core/components/tasks/optimized-tasks-components.tsx
@@ -1,3 +1,4 @@
+import { ActivityCalendarSkeleton } from '../common/skeleton/activity-calendar-skeleton';
 import { TaskCardSkeleton } from '../common/skeleton/profile-component-skeletons';
 import dynamic from 'next/dynamic';
 
@@ -6,5 +7,12 @@ export const LazyTaskCard = dynamic(
 	{
 		ssr: false,
 		loading: () => <TaskCardSkeleton />
+	}
+);
+export const LazyActivityCalendar = dynamic(
+	() => import('@/core/components/activities/activity-calendar').then((mod) => ({ default: mod.ActivityCalendar })),
+	{
+		ssr: false,
+		loading: () => <ActivityCalendarSkeleton />
 	}
 );
